Initial commit
Forgot to commit the first day.
Got the main window set up for the most part.
Known keys are being listed in the key manager window correctly.
Keys can be added and saved to a file for future use.
Still need to be able to delete keys and then get the main windows dropdown for key selection working properly.
After that, need to actually do the encrypt/decrypt logic.