Finished the encryption and decryption logic
Swapped to using the crypto library since it seems easier to use and has both AES and RSA support.  Had to make a few changes to how I was working the keys as a result.
Each file is encrypted with a unique AES key, which is also encrypted and saved with the key.
Selecting a directory recursively encrypts or decrypts all files within.
I'd like to set up something to hash the file and directory names next, still trying to decide the best way to accomplish this.